+++ docs/js/docs.js
@@ -231,7 +231,6 @@ docsApp.controller.DocsController = function($scope, $location, $window, section
       GLOBALS = /^angular\.([^\.]+)$/,
       MODULE = /^([^\.]+)$/,
       MODULE_MOCK = /^angular\.mock\.([^\.]+)$/,
-      MODULE_COMPONENT = /^(.+)\.components?:([^\.]+)$/,
       MODULE_CONTROLLER = /^(.+)\.controllers?:([^\.]+)$/,
       MODULE_DIRECTIVE = /^(.+)\.directives?:([^\.]+)$/,
       MODULE_DIRECTIVE_INPUT = /^(.+)\.directives?:input\.([^\.]+)$/,
@@ -251,8 +250,8 @@ docsApp.controller.DocsController = function($scope, $location, $window, section
       last: this.$last,
       active: page1 && this.currentPage == page1 || page2 && this.currentPage == page2,
       match: this.focused && this.currentPage != page1 &&
-             this.bestMatch.rank > 0 && this.bestMatch.page == page1,
-      deprecate: page1.isDeprecated
+             this.bestMatch.rank > 0 && this.bestMatch.page == page1
+
     };
   };
 
@@ -387,6 +386,11 @@ docsApp.controller.DocsController = function($scope, $location, $window, section
         search = $scope.search,
         bestMatch = {page: null, rank:0};
 
+    // Make sure the widgets are visible on every page's navigation
+    if ($location.path().split('/')[1] !== 'api') {
+      pages = pages.concat(sections.api);
+    }
+
     angular.forEach(pages, function(page) {
       var match,
         id = page.id,
@@ -400,7 +404,7 @@ docsApp.controller.DocsController = function($scope, $location, $window, section
 
       if (page.id == 'index') {
         //skip
-      } else if (!NG_DOCS.apis[section]) {
+      } else if (!NG_DOCS.apis[section] && page.url.indexOf('/api/') === -1) {
         otherPages.push(page);
       } else if (id == 'angular.Module') {
         module('ng', section).types.push(page);
@@ -410,8 +414,6 @@ docsApp.controller.DocsController = function($scope, $location, $window, section
         module(page.moduleName || match[1], section);
       } else if (match = id.match(MODULE_FILTER)) {
         module(page.moduleName || match[1], section).filters.push(page);
-      } else if (match = id.match(MODULE_COMPONENT)) {
-        module(page.moduleName || match[1], section).components.push(page);
       } else if (match = id.match(MODULE_CONTROLLER) && page.type === 'controller') {
         module(page.moduleName || match[1], section).controllers.push(page);
       } else if (match = id.match(MODULE_DIRECTIVE)) {
@@ -456,7 +458,6 @@ docsApp.controller.DocsController = function($scope, $location, $window, section
           name: name,
           url: (NG_DOCS.html5Mode ? '' : '#/') + section + '/' + name,
           globals: [],
-          components: [],
           controllers: [],
           directives: [],
           services: [],