+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This log tries to follow the good principles of [Keep a CHANGELOG](http://keepachangelog.com/).
 
+## 1.4.4 - 2020-03-25
+### Added
+- `odsChartQuery`: all parameters are now dynamic, meaning that if they are set to a variable, and the variable
+  changes, the chart will refresh accordingly.
+
 ## 1.4.3 - 2020-02-27
 ### Added
 - A new `fromjson` filter, to parse a text containing JSON and obtain the object